% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\name{EcologicalInferenceStrategyCalvo}
\alias{EcologicalInferenceStrategyCalvo}
\title{EcologicalInferenceStrategyCalvo}
\description{
EcologicalInferenceStrategyCalvo

EcologicalInferenceStrategyCalvo
}
\section{Super class}{
\code{\link[ElectionsLATAM:EcologicalInferenceStrategy]{ElectionsLATAM::EcologicalInferenceStrategy}} -> \code{EcologicalInferenceStrategyCalvo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EcologicalInferenceStrategyCalvo$new()}}
\item \href{#method-callDifp}{\code{EcologicalInferenceStrategyCalvo$callDifp()}}
\item \href{#method-paramsEstim}{\code{EcologicalInferenceStrategyCalvo$paramsEstim()}}
\item \href{#method-calcFractions}{\code{EcologicalInferenceStrategyCalvo$calcFractions()}}
\item \href{#method-paramsBoot}{\code{EcologicalInferenceStrategyCalvo$paramsBoot()}}
\item \href{#method-runEcologicalInference}{\code{EcologicalInferenceStrategyCalvo$runEcologicalInference()}}
\item \href{#method-clone}{\code{EcologicalInferenceStrategyCalvo$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ElectionsLATAM" data-topic="EcologicalInferenceStrategy" data-id="setProcessor">}\href{../../ElectionsLATAM/html/EcologicalInferenceStrategy.html#method-setProcessor}{\code{ElectionsLATAM::EcologicalInferenceStrategy$setProcessor()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$new(seed = 143324)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-callDifp"></a>}}
\if{latex}{\out{\hypertarget{method-callDifp}{}}}
\subsection{Method \code{callDifp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$callDifp(p, mx, my, covar, nR, nC, nP, const)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paramsEstim"></a>}}
\if{latex}{\out{\hypertarget{method-paramsEstim}{}}}
\subsection{Method \code{paramsEstim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$paramsEstim(
  data,
  x = -1,
  nR,
  nC,
  const = 0.001,
  parSeed = -1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcFractions"></a>}}
\if{latex}{\out{\hypertarget{method-calcFractions}{}}}
\subsection{Method \code{calcFractions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$calcFractions(p, nR, nC, covar = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paramsBoot"></a>}}
\if{latex}{\out{\hypertarget{method-paramsBoot}{}}}
\subsection{Method \code{paramsBoot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$paramsBoot(data, nR, nC, bootSamples)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runEcologicalInference"></a>}}
\if{latex}{\out{\hypertarget{method-runEcologicalInference}{}}}
\subsection{Method \code{runEcologicalInference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$runEcologicalInference(
  input.shares.fields,
  output.shares.fields
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyCalvo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
