% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ElectionsLATAM}
\alias{ElectionsLATAM}
\alias{.onLoad}
\title{ElectionsLATAM}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
A package to work with SQL datasources in a simple manner
}
\author{
Alejandro Baranek <abaranek@dc.uba.ar>
Executes code while loading the package.
}
