% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getEnv}
\alias{getEnv}
\title{getEnv}
\usage{
getEnv(
  variable.name,
  package.prefix = getPackagePrefix(),
  fail.on.empty = TRUE,
  env.file = "~/.Renviron",
  call.counter = 0,
  refresh.env = FALSE,
  logger = lgr
)
}
\arguments{
\item{variable.name}{the variable to be retrieved}

\item{package.prefix}{package prefix when retrieving from .Renviron (default getPackagePrefix())}

\item{fail.on.empty}{if an error should be returned when variable not found (default TRUE)}

\item{env.file}{environment file (default "~/.Renviron")}

\item{call.counter}{Internal use (default 0)}

\item{refresh.env}{refresh environment before retrieving (default FALSE)}

\item{logger}{for inheriting logger threshold from caller (default lgr)}
}
\description{
Returns the value of a variable name in .Renviron or in project .env file
}
\examples{
getEnv("variable", fail.on.empty = FALSE)
}
\author{
ken4rab
}
