% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\name{loadPivotInput}
\alias{loadPivotInput}
\title{loadPivotInput}
\usage{
loadPivotInput(input.filepath, col_types = cols(.default = col_number()))
}
\description{
loadPivotInput
}
\examples{
loadPivotInput(file.path(costa.rica.ein.path, "2021-generales_pivot_candidatos_n4.csv"))
}
\author{
ken4rab
}
