% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\name{loadPivotTable}
\alias{loadPivotTable}
\title{loadPivotTable}
\usage{
loadPivotTable(input.filepath, col_types = cols(.default = col_number()))
}
\arguments{
\item{input.filepath}{filepath for loading table}

\item{col_types}{readr col_types spec}
}
\description{
loadPivotTable
}
\examples{
costa.rica.ein.path <- file.path(getPackageDir(), "costa-rica")
loadPivotTable(file.path(costa.rica.ein.path, "2021-generales_pivot_candidatos_n4.csv"))
}
\author{
ken4rab
}
