% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loggerSetupFile}
\alias{loggerSetupFile}
\title{loggerSetupFile}
\usage{
loggerSetupFile(log.file)
}
\description{
Setup logger filename with a LayoutFormat
}
\examples{
log.filepath <- file.path(tempfile(), "lgr.log")
dir.create(tempfile(), recursive = TRUE, showWarnings = FALSE)
loggerSetupFile(log.filepath)

}
\author{
ken4rab
}
