% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\name{EcologicalInferenceStrategy}
\alias{EcologicalInferenceStrategy}
\title{EcologicalInferenceStrategy}
\description{
EcologicalInferenceStrategy

EcologicalInferenceStrategy
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EcologicalInferenceStrategy$new()}}
\item \href{#method-setProcessor}{\code{EcologicalInferenceStrategy$setProcessor()}}
\item \href{#method-clone}{\code{EcologicalInferenceStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$new(seed = 143324)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-setProcessor}{}}}
\subsection{Method \code{setProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$setProcessor(processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
