% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\docType{class}
\name{EcologicalInferenceStrategy}
\alias{EcologicalInferenceStrategy}
\title{EcologicalInferenceStrategy}
\description{
EcologicalInferenceStrategy

EcologicalInferenceStrategy
}
\examples{
ein <- EcologicalInferenceStrategy$new()
# Cannot execute an abstract class
# ein$runEcologicalInference(NULL, NULL)
}
\author{
ken4rab
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{seed}}{for initializing random generator}

\item{\code{processor}}{eir processor}

\item{\code{logger}}{lgr configured for class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EcologicalInferenceStrategy-new}{\code{EcologicalInferenceStrategy$new()}}
\item \href{#method-EcologicalInferenceStrategy-setProcessor}{\code{EcologicalInferenceStrategy$setProcessor()}}
\item \href{#method-EcologicalInferenceStrategy-runEcologicalInference}{\code{EcologicalInferenceStrategy$runEcologicalInference()}}
\item \href{#method-EcologicalInferenceStrategy-clone}{\code{EcologicalInferenceStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$new(seed = 143324)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategy-setProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategy-setProcessor}{}}}
\subsection{Method \code{setProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$setProcessor(processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategy-runEcologicalInference"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategy-runEcologicalInference}{}}}
\subsection{Method \code{runEcologicalInference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$runEcologicalInference(
  input.shares.fields,
  output.shares.fields
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
