% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological-inference.R
\docType{class}
\name{EcologicalInferenceStrategyWittenbergEtAl}
\alias{EcologicalInferenceStrategyWittenbergEtAl}
\title{EcologicalInferenceStrategyWittenbergEtAl}
\description{
EcologicalInferenceStrategyWittenbergEtAl

EcologicalInferenceStrategyWittenbergEtAl
}
\examples{
ein <- EcologicalInferenceStrategyWittenbergEtAl$new()
# Cannot run without having a processor
# ein$runEcologicalInference(NULL, NULL)
}
\author{
wittenberg
}
\section{Super class}{
\code{\link[ElectionsLATAM:EcologicalInferenceStrategy]{ElectionsLATAM::EcologicalInferenceStrategy}} -> \code{EcologicalInferenceStrategyWittenbergEtAl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{estsPG}}{estimation parameters}

\item{\code{input}}{shares}

\item{\code{dsOUTpre}}{shares}

\item{\code{fracsPG}}{fractions}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{input}}{shares}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-new}{\code{EcologicalInferenceStrategyWittenbergEtAl$new()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-callDifp}{\code{EcologicalInferenceStrategyWittenbergEtAl$callDifp()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-paramsEstim}{\code{EcologicalInferenceStrategyWittenbergEtAl$paramsEstim()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-calcFractions}{\code{EcologicalInferenceStrategyWittenbergEtAl$calcFractions()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-paramsBoot}{\code{EcologicalInferenceStrategyWittenbergEtAl$paramsBoot()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-runEcologicalInference}{\code{EcologicalInferenceStrategyWittenbergEtAl$runEcologicalInference()}}
\item \href{#method-EcologicalInferenceStrategyWittenbergEtAl-clone}{\code{EcologicalInferenceStrategyWittenbergEtAl$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ElectionsLATAM" data-topic="EcologicalInferenceStrategy" data-id="setProcessor"><a href='../../ElectionsLATAM/html/EcologicalInferenceStrategy.html#method-EcologicalInferenceStrategy-setProcessor'><code>ElectionsLATAM::EcologicalInferenceStrategy$setProcessor()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-new"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$new(seed = 143324)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-callDifp"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-callDifp}{}}}
\subsection{Method \code{callDifp()}}{
CALL.DIFP
Calculates penalty for given parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$callDifp(
  p,
  mx,
  my,
  covar,
  nR,
  nC,
  nP,
  const,
  w = 1/nP
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{- parameter vector R x (C-1)}

\item{\code{mx}}{- Column marginals}

\item{\code{my}}{- row marginals}

\item{\code{nR}}{- number of rows}

\item{\code{nC}}{- number of columns}

\item{\code{nP}}{- number of precincts}

\item{\code{const}}{- weight for penalty}

\item{\code{w}}{- weights for each observation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-paramsEstim"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-paramsEstim}{}}}
\subsection{Method \code{paramsEstim()}}{
Ecological Inference in the RxC case
Penalized Least Square Minimizer
PARAMS.ESTIM
Estimates parameters minimizing the penalized least squares criterion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$paramsEstim(
  data,
  x = -1,
  nR,
  nC,
  const = 0.001,
  parSeed = -1,
  w = 1/nrow(data)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{- marginals (optionally with covariates)}

\item{\code{x}}{- index (optional, for bootstrapping)}

\item{\code{nR}}{- number of rows}

\item{\code{nC}}{- number of columns}

\item{\code{const}}{- weight for penalty}

\item{\code{parSeed}}{- Seed for parameters (optional)}

\item{\code{w}}{- weights for each observation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-calcFractions"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-calcFractions}{}}}
\subsection{Method \code{calcFractions()}}{
Calculate Fractions
CALC.FRACTIONS
Calculate fractions from the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$calcFractions(p, nR, nC, covar = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{- parameters}

\item{\code{nR}}{- number of rows}

\item{\code{nC}}{- number of columns}

\item{\code{covar}}{- (Optional) Vector of covariates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-paramsBoot"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-paramsBoot}{}}}
\subsection{Method \code{paramsBoot()}}{
Bootstrapping
PARAMS.BOOT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$paramsBoot(data, nR, nC, bootSamples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{- marginals (optionally, with covariates)}

\item{\code{nR}}{- number of rows}

\item{\code{nC}}{- number of columns}

\item{\code{bootSamples}}{- number of bootstrap samples}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-runEcologicalInference"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-runEcologicalInference}{}}}
\subsection{Method \code{runEcologicalInference()}}{
runEcologicalInference
run ecological inference with current strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$runEcologicalInference(
  input.shares.fields,
  output.shares.fields,
  include.ausentes = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input.shares.fields}}{shares fields in input table}

\item{\code{output.shares.fields}}{shares fields in output table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcologicalInferenceStrategyWittenbergEtAl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EcologicalInferenceStrategyWittenbergEtAl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceStrategyWittenbergEtAl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
