% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getEnv}
\alias{getEnv}
\title{getEnv}
\usage{
getEnv(
  variable.name,
  package.prefix = getPackagePrefix(),
  fail.on.empty = TRUE,
  env.file = "~/.Renviron",
  call.counter = 0,
  refresh.env = FALSE,
  logger = lgr
)
}
\arguments{
\item{package.prefix}{= getPackagePrefix(),}

\item{fail.on.empty}{= TRUE}

\item{env.file}{= "~/.Renviron",}

\item{call.counter}{= 0,}

\item{refresh.env}{= FALSE,}

\item{logger}{= lgr}
}
\description{
Returns the value of a variable name in .Renviron or in project .env file
}
\examples{
getEnv("variable", fail.on.empty = FALSE)
}
\author{
ken4rab
}
