% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loggerSetupFile}
\alias{loggerSetupFile}
\title{loggerSetupFile}
\usage{
loggerSetupFile(log.filepath)
}
\arguments{
\item{log.filepath}{file path for setting a lgr::AppenderFile}
}
\description{
Setup logger filename with a LayoutFormat
}
\examples{
log.filepath <- file.path(tempdir(), "lgr.log")
dir.create(tempdir(), recursive = TRUE, showWarnings = FALSE)
loggerSetupFile(log.filepath)

}
\author{
ken4rab
}
